Partial fix for iDoubs issue 55

+++ branches/2.0/doubango/tinyDAV/src/audio/coreaudio/tdav_consumer_audiounit.c
@@ -148,7 +148,7 @@ static int tdav_consumer_audiounit_prepare(tmedia_consumer_t* self, const tmedia
 									  sizeof(audioFormat));
 		
 		if(status){
-			TSK_DEBUG_ERROR("AudioUnitSetProperty(kAudioUnitProperty_StreamFormat) failed with status=%d", (int32_t)status);
+			TSK_DEBUG_ERROR("AudioUnitSetProperty(kAudioUnitProperty_StreamFormat) failed with status=%ld", status);
 			return -5;
 		}
 		else {
@@ -163,7 +163,7 @@ static int tdav_consumer_audiounit_prepare(tmedia_consumer_t* self, const tmedia
 										  &callback, 
 										  sizeof(callback));
 			if(status){
-				TSK_DEBUG_ERROR("AudioUnitSetProperty(kAudioOutputUnitProperty_SetInputCallback) failed with status=%d", (int32_t)status);
+				TSK_DEBUG_ERROR("AudioUnitSetProperty(kAudioOutputUnitProperty_SetInputCallback) failed with status=%ld", status);
 				return -6;
 			}
 		}
